+++ src/Haskellorls/Config/Tree.hs
@@ -1,6 +1,3 @@
 module Haskellorls.Config.Tree (TreeNodePosition (..)) where
 
-data TreeNodePosition
-  = HEAD
-  | MID
-  | LAST
+data TreeNodePosition = HEAD | MID | LAST